+++ tests/Unit/PaymentManagerTest.php
@@ -55,12 +55,32 @@ class PaymentManagerTest extends TestCase
     public function a_payment_facade_invalid_exception_if_method_not_found(): void
     {
         $this->expectException(\InvalidArgumentException::class);
-        
+
         $method_id = 0;
         $payment = Payment::store($method_id);
     }
     /** @test */
     public function it_can_set_amount_when_add_to_items()
+    {
+        $payment = Payment::store(2);
+        $items = [
+            "name" => "ASC1515",
+            "amount_cents" => 500000,
+            "description" => "Smart Watch",
+            "quantity" => "1"
+        ];
+        $payment->items($items);
+
+        $reflector = new \ReflectionClass(PaymobMethod::class);
+        $property = $reflector->getProperty('amount');
+        $property->setAccessible(true);
+
+        $amount = $property->getValue($payment);
+
+        $this->assertEquals($items['amount_cents'], $amount);
+    }
+    /** @test */
+    public function it_can_add_to_items_without_amount()
     {
         $payment = Payment::store(2);
         $items = [