+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## [Unreleased](https://github.com/shabayekdes/laravel-payment/compare/main...develop)
 
+## [v0.5.1 (2022-01-07)](https://github.com/shabayekdes/laravel-payment/compare/v0.5.0...v0.5.1)
+
+### Added
+- Add credentails paymob property ([#2](https://github.com/shabayekdes/laravel-payment/pull/2))
+- Add github action workflow ([#2](https://github.com/shabayekdes/laravel-payment/pull/2))
+
+### Changed
+- Convert payment to singleton class and add payment facade aliase ([#2](https://github.com/shabayekdes/laravel-payment/pull/2))
 
 ## [v0.5 (2022-01-07)](https://github.com/shabayekdes/laravel-payment/compare/v0.4.1...develop)
 
 ### Changed
-- Set order total amount when add items ([#5750](https://github.com/shabayekdes/laravel-payment/pull/1))
+- Set order total amount when add items ([#1](https://github.com/shabayekdes/laravel-payment/pull/1))